Refactored time-related functions into time.c
