+++ time.h
@@ -1,0 +1,9 @@
+#ifndef TIME_H_INCLUDED
+#define TIME_H_INCLUDED
+
+int day_of_the_year(int day, int month, int year);
+int isLeapyear(int year);
+int get_days_for_month(int month, int year);
+int exists_date(int day, int month, int year);
+
+#endif // TIME_H_INCLUDED
